+++ svg-to-compose/src/commonMain/kotlin/dev/tonholo/s2c/geom/path/PathInstructionToAbsolute.kt
@@ -31,76 +31,76 @@ fun List<PathNodes>.toAbsolute(): List<PathNodes> {
 
 private fun PathNodes.LineTo.toAbsolute(current: DoubleArray): PathNodes.LineTo =
     if (isRelative.not()) {
-        current[0] = x.toDouble()
-        current[1] = y.toDouble()
+        current[0] = x
+        current[1] = y
         this
     } else {
-        current[0] += x.toDouble()
-        current[1] += y.toDouble()
+        current[0] += x
+        current[1] += y
         copy(
-            x = current[0].toFloat(),
-            y = current[1].toFloat(),
+            x = current[0],
+            y = current[1],
             isRelative = false,
         )
     }
 
 private fun PathNodes.MoveTo.toAbsolute(current: DoubleArray): PathNodes.MoveTo =
     if (isRelative.not()) {
-        current[0] = x.toDouble()
-        current[1] = y.toDouble()
+        current[0] = x
+        current[1] = y
         this
     } else {
-        current[0] += x.toDouble()
-        current[1] += y.toDouble()
+        current[0] += x
+        current[1] += y
         copy(
-            x = current[0].toFloat(),
-            y = current[1].toFloat(),
+            x = current[0],
+            y = current[1],
             isRelative = false,
         )
     }
 
 private fun PathNodes.HorizontalLineTo.toAbsolute(current: DoubleArray): PathNodes.HorizontalLineTo =
     if (isRelative.not()) {
-        current[0] = x.toDouble()
+        current[0] = x
         this
     } else {
-        current[0] += x.toDouble()
+        current[0] += x
         copy(
-            x = current[0].toFloat(),
+            x = current[0],
             isRelative = false,
         )
     }
 
 private fun PathNodes.VerticalLineTo.toAbsolute(current: DoubleArray): PathNodes.VerticalLineTo =
     if (isRelative.not()) {
-        current[1] = y.toDouble()
+        current[1] = y
         this
     } else {
-        current[1] += y.toDouble()
+        current[1] += y
         copy(
-            y = current[1].toFloat(),
+            y = current[1],
             isRelative = false,
         )
     }
 
 fun PathNodes.CurveTo.toAbsolute(current: DoubleArray): PathNodes.CurveTo =
     if (isRelative.not()) {
-        current[0] = x3.toDouble()
-        current[1] = y3.toDouble()
+        current[0] = x3
+        current[1] = y3
         this
     } else {
         val currentX = current[0]
         val currentY = current[1]
         copy(
-            x1 = currentX.toFloat() + x1,
-            y1 = currentY.toFloat() + y1,
-            x2 = currentX.toFloat() + x2,
-            y2 = currentY.toFloat() + y2,
-            x3 = (currentX.toFloat() + x3).also {
-                current[0] = it.toDouble()
+            x1 = currentX + x1,
+            y1 = currentY + y1,
+            x2 = currentX + x2,
+            y2 = currentY + y2,
+            x3 = (currentX + x3).also {
+                current[0] = it
             },
-            y3 = (currentY.toFloat() + y3).also {
-                current[1] = it.toDouble()
+            y3 = (currentY + y3).also {
+                current[1] = it
             },
             isRelative = false,
         )
@@ -108,20 +108,20 @@ fun PathNodes.CurveTo.toAbsolute(current: DoubleArray): PathNodes.CurveTo =
 
 fun PathNodes.ReflectiveCurveTo.toAbsolute(current: DoubleArray): PathNodes.ReflectiveCurveTo =
     if (isRelative.not()) {
-        current[0] = x2.toDouble()
-        current[1] = y2.toDouble()
+        current[0] = x2
+        current[1] = y2
         this
     } else {
         val currentX = current[0]
         val currentY = current[1]
         copy(
-            x1 = currentX.toFloat() + x1,
-            y1 = currentY.toFloat() + y1,
-            x2 = (currentX.toFloat() + x2).also {
-                current[0] = it.toDouble()
+            x1 = currentX + x1,
+            y1 = currentY + y1,
+            x2 = (currentX + x2).also {
+                current[0] = it
             },
-            y2 = (currentY.toFloat() + y2).also {
-                current[1] = it.toDouble()
+            y2 = (currentY + y2).also {
+                current[1] = it
             },
             isRelative = false,
         )
@@ -129,20 +129,20 @@ fun PathNodes.ReflectiveCurveTo.toAbsolute(current: DoubleArray): PathNodes.Refl
 
 private fun PathNodes.QuadTo.toAbsolute(current: DoubleArray): PathNodes.QuadTo =
     if (isRelative.not()) {
-        current[0] = x2.toDouble()
-        current[1] = y2.toDouble()
+        current[0] = x2
+        current[1] = y2
         this
     } else {
         val currentX = current[0]
         val currentY = current[1]
         copy(
-            x1 = currentX.toFloat() + x1,
-            y1 = currentY.toFloat() + y1,
-            x2 = (currentX.toFloat() + x2).also {
-                current[0] = it.toDouble()
+            x1 = currentX + x1,
+            y1 = currentY + y1,
+            x2 = (currentX + x2).also {
+                current[0] = it
             },
-            y2 = (currentY.toFloat() + y2).also {
-                current[1] = it.toDouble()
+            y2 = (currentY + y2).also {
+                current[1] = it
             },
             isRelative = false,
         )
@@ -150,30 +150,30 @@ private fun PathNodes.QuadTo.toAbsolute(current: DoubleArray): PathNodes.QuadTo 
 
 private fun PathNodes.ReflectiveQuadTo.toAbsolute(current: DoubleArray): PathNodes.ReflectiveQuadTo =
     if (isRelative.not()) {
-        current[0] = x1.toDouble()
-        current[1] = y1.toDouble()
+        current[0] = x1
+        current[1] = y1
         this
     } else {
-        current[0] += x1.toDouble()
-        current[1] += y1.toDouble()
+        current[0] += x1
+        current[1] += y1
         copy(
-            x1 = current[0].toFloat(),
-            y1 = current[1].toFloat(),
+            x1 = current[0],
+            y1 = current[1],
             isRelative = false,
         )
     }
 
 fun PathNodes.ArcTo.toAbsolute(current: DoubleArray): PathNodes.ArcTo =
     if (isRelative.not()) {
-        current[0] = x.toDouble()
-        current[1] = y.toDouble()
+        current[0] = x
+        current[1] = y
         this
     } else {
-        current[0] += x.toDouble()
-        current[1] += y.toDouble()
+        current[0] += x
+        current[1] += y
         copy(
-            x = current[0].toFloat(),
-            y = current[1].toFloat(),
+            x = current[0],
+            y = current[1],
             isRelative = false,
         )
     }
